refactor: add folder pages
