+++ src/app/pages/authorization.ts
@@ -1,4 +1,4 @@
-import Component from "../common/Component";
+import Component from "../../common/Component";
 
 interface IUserData {
   email: string,
