+++ src/app/pages/mainpage.ts
@@ -1,0 +1,1 @@
+import Component from "../../common/Component";
